Separating auth methods from base
